Day10/Practice/requirements.txt
kafka-python==2.0.2
web3==6.0.0
python-dotenv==1.0.0
confluent-kafka==2.1.1
pyyaml==6.0.1
requests==2.31.0 